# src/main/resources/api-specs/user-openapi.yaml
openapi: "3.0.3"

info:
  title: "Training Project User API"
  description: "API for accessing user profiles and related operations in the TrainingProject platform."
  version: "1.0.0"
servers:
  - url: "https://localhost:8083"
    description: "Local development server"
  - url: "https://training-project.uk/backend"
    description: "Production server"

security:
  - bearerAuth: []

tags:
  - name: "User Profile"
    description: "Operations related to user profile and account settings."
  - name: "Delivery Addresses"
    description: "Operations related to user delivery addresses."
  - name: "User Avatar"
    description: "Operations related to user avatars."

paths:
  /api/v1/users:

    get:
      tags:
        - "User Profile"
      summary: "Get user profile"
      description: "Fetch detailed information about the authenticated user."
      operationId: "getUserProfile"
      responses:
        "200":
          description: "A user object containing detailed user information."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "User not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - "User Profile"
      summary: "Edit user profile"
      description: "Update the authenticated user's profile details."
      operationId: "editUserProfile"
      requestBody:
        description: "Payload containing the user's updated details."
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserAccountRequest"
      responses:
        "200":
          description: "The updated user object."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          description: "Invalid input data."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "User not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - "User Profile"
      summary: "Change user password"
      description: "Change the password for the authenticated user."
      operationId: "changeUserPassword"
      requestBody:
        description: "Payload containing the old and new passwords."
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUserPasswordRequest"
      responses:
        "200":
          description: "Password changed successfully."
        "400":
          description: "Validation error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidPassword:
                  value: { "message": "Password must be at least 8 characters long and contain at least one lowercase letter, one uppercase letter, and one digit." }
        "401":
          description: "Unauthorized access or invalid old password."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - "User Profile"
      summary: "Delete user profile"
      description: "Delete the authenticated user's profile."
      operationId: "deleteUserProfile"
      responses:
        "200":
          description: "User successfully deleted."
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "User not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/addresses:
    get:
      tags:
        - "Delivery Addresses"
      summary: "Get delivery addresses"
      operationId: "getDeliveryAddresses"
      responses:
        "200":
          description: "List of delivery addresses."
          content:
            application/json:
              schema:
                title: "DeliveryAddressList"
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryAddressDto"
        "401":
          description: "Unauthorized."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - "Delivery Addresses"
      summary: "Add delivery address"
      operationId: "addDeliveryAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryAddressRequest"
      responses:
        "201":
          description: "Address created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryAddressDto"
        "400":
          description: "Invalid input."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Unauthorized."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/addresses/{addressId}:
    put:
      tags:
        - "Delivery Addresses"
      summary: "Update delivery address"
      operationId: "updateDeliveryAddress"
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryAddressRequest"
      responses:
        "200":
          description: "Address updated."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryAddressDto"
        "404":
          description: "Address not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - "Delivery Addresses"
      summary: "Delete delivery address"
      operationId: "deleteDeliveryAddress"
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: "Address deleted."
        "404":
          description: "Address not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/addresses/{addressId}/default:
    patch:
      tags:
        - "Delivery Addresses"
      summary: "Set default delivery address"
      operationId: "setDefaultDeliveryAddress"
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Default address updated."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryAddressDto"
        "404":
          description: "Address not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/avatar:
    post:
      tags:
        - "User Avatar"
      summary: "Upload user avatar"
      description: "Upload a user's avatar image."
      operationId: "uploadUserAvatar"
      requestBody:
        description: "Payload containing the avatar image file."
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadUserAvatarRequest"
      responses:
        "200":
          description: "User avatar uploaded successfully."
        "400":
          description: "Invalid file format or size."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: "Avatar upload is temporarily unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "User not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - "User Avatar"
      summary: "Get user avatar link"
      description: "Retrieve the link to the authenticated user's avatar image."
      operationId: "getUserAvatarLink"
      responses:
        "200":
          description: "User avatar link retrieved successfully."
          content:
            application/json:
              schema:
                title: UserAvatarLink
                type: "string"
        "404":
          description: "User has no avatar."
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - "User Avatar"
      summary: "Delete user avatar"
      description: "Delete the authenticated user's avatar image."
      operationId: "deleteUserAvatar"
      responses:
        "200":
          description: "User avatar deleted successfully."
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/avatar/uploads:
    post:
      tags:
        - "User Avatar"
      summary: "Create avatar upload intent"
      description: "Create a short-lived avatar upload intent for direct storage upload. Disabled unless presigned avatar upload mode is configured."
      operationId: "createAvatarUpload"
      parameters:
        - name: "Idempotency-Key"
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvatarUploadRequest"
      responses:
        "200":
          description: "Avatar upload intent created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadIntentResponse"
        "400":
          description: "Invalid request or expired idempotency retry."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: "Presigned avatar upload is not configured."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/avatar/uploads/{uploadId}:
    get:
      tags:
        - "User Avatar"
      summary: "Get avatar upload status"
      description: "Get the current status for an avatar upload owned by the authenticated user."
      operationId: "getAvatarUpload"
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Avatar upload status."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadStatusResponse"
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Upload not found."
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - "User Avatar"
      summary: "Cancel avatar upload"
      description: "Cancel the authenticated user's in-flight or pending avatar upload intent. Idempotent and safe to retry."
      operationId: "cancelAvatarUpload"
      parameters:
        - name: "uploadId"
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: "Avatar upload canceled or already absent."
        "401":
          description: "Unauthorized access. Authentication required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token for authenticated requests"

  schemas:
    UploadUserAvatarRequest:
      title: UploadUserAvatarRequest
      type: "object"
      required:
        - file
      properties:
        file:
          title: UserAvatarFile
          type: "string"
          format: "binary"
        turnstileToken:
          type: "string"
          maxLength: 2048
          description: "Cloudflare Turnstile verification token. Required when avatar upload bot protection is enabled."

    CreateAvatarUploadRequest:
      title: CreateAvatarUploadRequest
      type: object
      required:
        - contentType
        - sizeBytes
      properties:
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
        turnstileToken:
          type: string
          maxLength: 2048
          description: "Cloudflare Turnstile verification token. Required when avatar upload bot protection is enabled."

    AvatarUploadStatus:
      type: string
      enum:
        - PENDING_UPLOAD
        - UPLOADED
        - PROCESSING
        - READY
        - FAILED
        - EXPIRED
        - SUPERSEDED

    AvatarUploadTargetResponse:
      type: object
      nullable: true
      properties:
        method:
          type: string
          enum:
            - POST
            - PUT
        url:
          type: string
          format: uri
        fields:
          type: object
          additionalProperties:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string

    AvatarUploadIntentResponse:
      type: object
      required:
        - uploadId
        - status
        - expiresAt
      properties:
        uploadId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AvatarUploadStatus"
        upload:
          $ref: "#/components/schemas/AvatarUploadTargetResponse"
        expiresAt:
          type: string
          format: date-time

    AvatarUploadStatusResponse:
      type: object
      required:
        - uploadId
        - status
        - expiresAt
      properties:
        uploadId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AvatarUploadStatus"
        avatarLink:
          type: string
          nullable: true
        failureCode:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time

    UpdateUserAccountRequest:
      type: "object"
      description: "Payload containing the user's updated details."
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: "string"
          description: "The first name of the user."
          minLength: 2
          maxLength: 64
          pattern: "^[a-zA-Z\\u00C0-\\u00D6\\u00D8-\\u00F6\\u00F8-\\u00FF\\s''\\u2019\\-]+$"
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message = \"First name is required\")"
          example: "John"
        lastName:
          type: "string"
          description: "The last name of the user."
          minLength: 2
          maxLength: 64
          pattern: "^[a-zA-Z\\u00C0-\\u00D6\\u00D8-\\u00F6\\u00F8-\\u00FF\\s''\\u2019\\-]+$"
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message = \"Last name is required\")"
          example: "Doe"
        birthDate:
          type: "string"
          description: "The birth date of the user."
          format: date
        phoneNumber:
          type: "string"
          description: "The phone number of the user."
          example: "+1234567890"
        address:
          $ref: "#/components/schemas/AddressDto"

    PasswordField:
      type: "string"
      description: "Password must be between 8 and 128 characters, with at least one lowercase letter, one uppercase letter, and one digit."
      minLength: 8
      maxLength: 128
      pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$"

    ChangeUserPasswordRequest:
      type: "object"
      description: "Payload for changing user password."
      required:
        - newPassword
        - oldPassword
      properties:
        newPassword:
          $ref: "#/components/schemas/PasswordField"
        oldPassword:
          $ref: "#/components/schemas/PasswordField"

    UserDto:
      type: "object"
      description: "A user profile in the TrainingProject system."
      required:
        - firstName
        - lastName
        - email
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "The unique identifier for the user."
        firstName:
          type: "string"
          minLength: 2
          maxLength: 64
          description: "The first name of the user."
        lastName:
          type: "string"
          minLength: 2
          maxLength: 64
          description: "The last name of the user."
        birthDate:
          type: "string"
          description: "The birth date of the user."
          format: date
        phoneNumber:
          type: "string"
          description: "The phone number of the user."
        stripeCustomerToken:
          type: "string"
          description: "The Stripe customer token associated with the user."
        email:
          type: "string"
          description: "The email address of the user."
          format: "email"
        address:
          $ref: "#/components/schemas/AddressDto"
        avatarLink:
          type: "string"
          description: "The avatar URL of the user."
        oauthUser:
          type: "boolean"
          description: "True if the user signed up via OAuth and has no password set."

    AddressDto:
      type: "object"
      description: "Address details of the user."
      properties:
        country:
          type: "string"
          minLength: 2
          maxLength: 128
          description: "The country of the address."
          example: "United Kingdom"
        city:
          type: "string"
          minLength: 2
          maxLength: 128
          description: "The city of the address."
          example: "London"
        line:
          type: "string"
          minLength: 2
          maxLength: 128
          description: "The address line."
          example: "123 Coffee St."
        postcode:
          type: "string"
          minLength: 2
          maxLength: 128
          description: "The postal code of the address."
          example: "E1 6AN"

    InitiatePasswordResetRequest:
      type: "object"
      description: "Payload to initiate a password reset."
      required:
        - email
      properties:
        email:
          type: "string"
          format: "email"
          description: "The email address of the user requesting a password reset."

    ConfirmPasswordResetRequest:
      type: "object"
      description: "Payload to confirm a password reset."
      required:
        - token
        - newPassword
      properties:
        token:
          type: "string"
          description: "The password reset token."
        newPassword:
          $ref: "#/components/schemas/PasswordField"

    DeliveryAddressDto:
      type: "object"
      description: "A saved delivery address."
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          example: "Home"
        line:
          type: string
          example: "123 Coffee St"
        city:
          type: string
          example: "London"
        country:
          type: string
          example: "United Kingdom"
        postcode:
          type: string
          example: "E1 6AN"
        isDefault:
          type: boolean

    DeliveryAddressRequest:
      type: "object"
      required:
        - label
        - line
        - city
        - country
        - postcode
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 64
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"Address label must not be blank\")"
          example: "Home"
        line:
          type: string
          minLength: 1
          maxLength: 256
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"Address line must not be blank\")"
          example: "123 Coffee St"
        city:
          type: string
          minLength: 1
          maxLength: 128
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"Address city must not be blank\")"
          example: "London"
        country:
          type: string
          minLength: 1
          maxLength: 128
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"Address country must not be blank\")"
          example: "United Kingdom"
        postcode:
          type: string
          minLength: 1
          maxLength: 16
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message = \"Address postcode must not be blank\")"
          example: "E1 6AN"

    ErrorResponse:
      $ref: "common-schemas.yaml#/components/schemas/ErrorResponse"